+++ mitra_models/migrations/V0117__payment_method__add_view_key.sql
@@ -1,0 +1,1 @@
+ALTER TABLE payment_method ADD COLUMN view_key VARCHAR(500);