+++ mitra_services/src/monero/mod.rs
@@ -1,3 +1,4 @@
 pub mod caip122;
+pub mod light_wallet;
 pub mod utils;
 pub mod wallet;